Refactor camera and scene-related functionality into dedicated `CameraUtility` and `CharacterScene` files.
