+++ src/source/Scenes/MainScene.h
@@ -1,0 +1,9 @@
+#pragma once
+
+// MainScene.h - Main game scene
+
+#include <windows.h>
+
+// Main scene lifecycle
+void MoveMainScene();
+bool RenderMainScene();
