+++ src/source/Scenes/WebzenScene.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// WebzenScene.h - Splash screen scene
+
+#include <windows.h>
+
+// Splash screen scene
+void WebzenScene(HDC hDC);
